+++ src/Http/Output/Html.php
@@ -21,8 +21,6 @@ class Html implements \Parable\Http\Output\OutputInterface
      */
     public function prepare(\Parable\Http\Response $response)
     {
-        if (!is_string($response->getContent()) && !is_null($response->getContent())) {
-            throw new \Parable\Http\Exception('Can only work with string or null content');
-        }
+        $response->setContent($response->getContentAsString());
     }
 }