+++ tests/Components/Http/Output/JsonTest.php
@@ -37,18 +37,6 @@ class JsonTest extends \Parable\Tests\Base
 
         $this->json->prepare($response);
 
-        $this->assertSame('{"value":"stuff","secondary":"what now"}', $response->getContent());
-    }
-
-    public function testInvalidJsonStringStaysString()
-    {
-        $response = \Parable\DI\Container::createAll(\Parable\Http\Response::class);
-        $response->setOutput($this->json);
-
-        $response->setContent("{[{[}]}]");
-
-        $this->json->prepare($response);
-
-        $this->assertSame('"{[{[}]}]"', $response->getContent());
+        $this->assertSame('{"value":"stuff","secondary":"what now"}', $response->getContentAsString());
     }
 }